refactor: create a directory for views
